deleted old relation file
